requirements.txt
# Bot kutubxonalari
aiogram==3.4.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Yordamchi kutubxonalar
aiohttp==3.9.1
pydantic==2.5.3

# To'lov tizimlari (ixtiyoriy)
# click-uz==0.1.0
# payme-uz==0.1.0

# Boshqa foydali kutubxonalar
python-dateutil==2.8.2